LevelOrderTraversal refactoring
